SSE-22 Closing SSE session for Ranger will inadvertently shut down OutingChannel for same user wi...
- Adds extension of EventOutput that holds the Request ID for tracking
of the UserChannel info within the OutingChannel's broadcaster.
- Adds further diagnostic logging to watch what EventOutput instances
are within the broadcaster.

+++ src/main/java/com/clueride/sse/channel/UserChannelServiceImpl.java
@@ -58,7 +58,10 @@ public class UserChannelServiceImpl implements UserChannelService {
         } else {
             originalRequestId = userChannel.getRequestId();
             userChannel.getKeepAliveGenerator().release();
-            outputChannelService.removeUserChannelFromOuting(userChannel);
+            if (userChannel.getOutingId() != null) {
+                /* Only attempt to remove this UserChannel if we have an outing. */
+                outputChannelService.removeUserChannelFromOuting(userChannel);
+            }
             channelPerChunkedOutput.remove(chunkedOutput);
         }
